Parametrized metrics pusher
