+++ core/serv-api/src/lib.rs
@@ -4,12 +4,19 @@ use serde::Serialize;
 use std::path::PathBuf;
 
 #[derive(Debug, Default)]
+pub struct MetricsCtx {
+    pub push_enabled: bool,
+    pub push_host_url: Option<url::Url>,
+}
+
+#[derive(Debug, Default)]
 pub struct CliCtx {
     pub data_dir: PathBuf,
     pub gsb_url: Option<url::Url>,
     pub json_output: bool,
     pub accept_terms: bool,
     pub interactive: bool,
+    pub metrics: MetricsCtx,
 }
 
 impl CliCtx {
